refactor(famiglie): Update route names and remove unused MatrimonioController
